anatomy was moved to openpype pipeline
